bridge-vlan: add wan port only behind swconfig
Configure only the switch port for the WAN port instead of all Ethernet
ports.  When adding a dynamic VLAN, it should be added to the WAN port
as tagged only, and not the LAN ports.

When LAN ports were included in the interface config, it resulted in
their switch ports also being added to event.swconfig.ports as well as
duplicating the CPU port for each LAN port.   This also caused the
swconfig command to fail.

Config before the fix:
    config config 'config'
        list wan_port 'eth0.1'
        option swconfig 'switch1'
        list swconfig_ports '5t'
        list swconfig_ports '6t'
        list swconfig_ports '2t'
        list swconfig_ports '6t'
        list swconfig_ports '3t'
        list swconfig_ports '6t'
        list swconfig_ports '4t'
        list swconfig_ports '6t'

Config after the fix:
    config config 'config'
        list wan_port 'eth0.1'
        option swconfig 'switch1'
        list swconfig_ports '5t'
        list swconfig_ports '6t'

Signed-off-by: Paul White <paul@shasta.cloud>