Fixing issues with section names in network config
In ethernet.uc, the section names are the same as that of interface names with . replaced by _
However name and ifname should still be the name with the . (dot). Fixed it.
Interface with _ in the name does n't exist

In ieee8021x.uc , auth and aut_vlan should be updated on the already created section in the network
and not on a new section. Otherwise a redundant device section would be created which is not correct.

Signed-off-by: Venkat Chimata <venkata@shasta.cloud>