handle multi-band phys for encryption check
Do not validate encryption against bands supported by the phy, but
against the band in use by getting the cross-section of ssid bands and
phy bands.

Fixes no default wireless configuration being generated if the phy is
both 5G and 6G capable, as all 6G encryption constraints are applied
even if the radio isn't configured for the 6G band.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>